Refactor authentication context and hooks; replace AuthContext with a new definition and implement useAuth hook for better context management. Update login form, navbar, and protected routes to utilize the new auth hooks. Add badge and navigation menu components with associated styles for improved UI consistency.
