+++ cliente_web/src/contexts/auth-context-definition.ts
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+export interface AuthContextType {
+  isAuthenticated: boolean;
+  isLoading: boolean;
+  login: (token: string) => void;
+  logout: () => void;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
