+++ cliente_web/src/contexts/auth-hooks.ts
@@ -1,0 +1,11 @@
+import { useContext } from "react"
+import { AuthContext } from "./auth-context-definition"
+
+export const useAuth = () => {
+  const context = useContext(AuthContext)
+  if (context === undefined) {
+    throw new Error("useAuth must be used within an AuthProvider")
+  }
+  return context
+}
+
